+++ src/ActivityRepositoryInterface.php
@@ -27,9 +27,9 @@ interface ActivityRepositoryInterface
      *
      * @param IRI $iri The activity id to filter by
      *
-     * @return Activity The activity
+     * @return Activity|null The activity
      *
      * @throws NotFoundException if no Activity with the given IRI does exist
      */
-    public function findActivityById(IRI $iri);
+    public function findActivityById(IRI $iri): ?Activity;
 }